+++ recipes/master_config_post.rb
@@ -124,32 +124,23 @@ execute 'Wait up to 30s for nodes registration' do
   retry_delay 5
 end
 
-if node['cookbook-openshift3']['openshift_hosted_manage_router']
-  execute 'Deploy Hosted Router' do
-    command "#{node['cookbook-openshift3']['openshift_common_client_binary']} adm router --selector=${selector_router} -n ${namespace_router} --config=admin.kubeconfig || true"
-    environment(
-      'selector_router' => node['cookbook-openshift3']['openshift_hosted_router_selector'],
-      'namespace_router' => node['cookbook-openshift3']['openshift_hosted_router_namespace']
-    )
-    cwd Chef::Config[:file_cache_path]
-    only_if '[[ `oc get pod --selector=router=router --config=admin.kubeconfig | wc -l` -eq 0 ]]'
-  end
-  execute 'Auto Scale Router based on label' do
-    command "#{node['cookbook-openshift3']['openshift_common_client_binary']} scale dc/router --replicas=${replica_number} -n ${namespace_router} --config=admin.kubeconfig"
-    environment(
-      'replica_number' => Mixlib::ShellOut.new("oc get node --no-headers --selector=#{node['cookbook-openshift3']['openshift_hosted_router_selector']} --config=#{Chef::Config[:file_cache_path]}/admin.kubeconfig | wc -l").run_command.stdout.strip,
-      'namespace_router' => node['cookbook-openshift3']['openshift_hosted_router_namespace']
-    )
-    cwd Chef::Config[:file_cache_path]
-    not_if '[[ `oc get pod --selector=router=router --config=admin.kubeconfig --no-headers | wc -l` -eq ${replica_number} ]]'
+openshift_deploy_router 'Deploy Router' do
+  only_if do
+    node['cookbook-openshift3']['openshift_hosted_manage_router']
   end
 end
 
-if node['cookbook-openshift3']['openshift_hosted_manage_registry']
-  openshift_deploy_registry 'Deploy Registry' do
-    persistent_registry node['cookbook-openshift3']['registry_persistent_volume'].empty? ? false : true
-    only_if do
-      node['cookbook-openshift3']['openshift_hosted_manage_registry']
-    end
+openshift_deploy_registry 'Deploy Registry' do
+  persistent_registry node['cookbook-openshift3']['registry_persistent_volume'].empty? ? false : true
+  persistent_volume_claim_name "#{node['cookbook-openshift3']['registry_persistent_volume']}-claim"
+  only_if do
+    node['cookbook-openshift3']['openshift_hosted_manage_registry']
+  end
+end
+
+openshift_deploy_metrics 'Deploy Cluster Metrics' do
+  metrics_params Hash[node['cookbook-openshift3']['openshift_hosted_metrics_parameters'].map { |k, v| [k.upcase, v] }]
+  only_if do
+    node['cookbook-openshift3']['openshift_hosted_cluster_metrics']
   end
 end