+++ recipes/validate.rb
@@ -16,3 +16,9 @@ end
 if node['cookbook-openshift3']['openshift_HA'] && node['cookbook-openshift3']['openshift_cluster_name'].nil?
   Chef::Application.fatal!('A Cluster Name must be defined via "openshift_cluster_name"')
 end
+
+if node['cookbook-openshift3']['openshift_hosted_cluster_metrics']
+  unless node['cookbook-openshift3']['openshift_hosted_cluster_metrics'] && Hash[node['cookbook-openshift3']['openshift_hosted_metrics_parameters'].map { |k, v| [k.upcase, v] }].key?('HAWKULAR_METRICS_HOSTNAME')
+    Chef::Application.fatal!('Key HAWKULAR_METRICS_HOSTNAME must be defined when deploying cluster metrics ("openshift_hosted_metrics_parameters")')
+  end
+end