+++ test/inspec/shared/12_directory_permissions_test.rb
@@ -1,0 +1,23 @@
+describe directory('/etc/origin') do
+  its('owner') { should eq 'root' }
+  its('group') { should eq 'root' }
+  its('mode') { should cmp '0755' }
+end
+
+describe directory('/etc/origin/master') do
+  its('owner') { should eq 'root' }
+  its('group') { should eq 'root' }
+  its('mode') { should cmp '0700' }
+end
+
+describe directory('/etc/origin/node') do
+  its('owner') { should eq 'root' }
+  its('group') { should eq 'root' }
+  its('mode') { should cmp '0755' }
+end
+
+describe directory('/etc/origin/node/ca.crt') do
+  its('owner') { should eq 'root' }
+  its('group') { should eq 'root' }
+  its('mode') { should cmp '0644' }
+end