+++ CHANGELOG.md
@@ -1,6 +1,49 @@
 # Openshift 3 Cookbook CHANGELOG
 This file is used to list changes made in each version of the Openshift 3 cookbook.
 
+## v1.10.30
+### Improvement
+- Add the possibility to deploy the cluster metrics
+- Add the possibility to add more manageName serviceaccount in master config
+- Move registry persistent_volume_claim name to explicit LWRP attribute
+- Added integration test for openshift_hosted_manage_registry feature
+- Added integration test for openshift_hosted_manage_router feature
+- Added integration test for persistent_storage feature
+- Refactor router-related resources to new openshift_deploy_router LWRP
+- Move registry persistent_volume_claim name to explicit LWRP attribute
+
+### Bug
+- Fix README.md typo
+- Fix issue with systemd when uninstalling the Openshift
+- Fix issue for systemctl daemon-reload
+- Removed redundant guard clause for registry deloyment
+
+## v1.10.29
+### Bug
+- Remove property attributes for resources (backward compatibility)
+
+## v1.10.28
+### Improvement
+- Add the possibility to deploy the cluster metrics
+- Add the possibility to add more manageName serviceaccount in master config
+
+### Bug
+- Fix README.md typo
+- Fix issue with systemd when uninstalling the Openshift
+- Fix issue for systemctl daemon-reload
+
+## v1.10.27
+### Bug
+- Fix the Origin deployment issue (https://github.com/IshentRas/cookbook-openshift3/issues/20)
+- Fix master-api service and master-controllers service (https://github.com/IshentRas/cookbook-openshift3/issues/40)
+
+## v1.10.26
+### Improvement
+- Set the default ipaddress used in etcd-related attributes accordingly with the etcd_server variable
+
+### Bug
+- Remove duplicated variables for ETCD
+
 ## v1.10.25
 ### Bug
 - Fix documentation
@@ -54,7 +97,7 @@ This file is used to list changes made in each version of the Openshift 3 cookbo
 - Add the possibility to enable the Audit logging 
 - Add the possibility to label nodes
 - Add the possibility to set scheduling against nodes
-- Add the plossibilty to deploy the Stand-alone Registry
+- Add the possibility to deploy the Stand-alone Registry & Router
 
 ### Bug
 - Remove automatic rebooting when playing adhoc uninstallation